Continue rename
